Missiles shooting, TODO: close the threads

+++ src/main/java/com/schalldach/thomas/game/helper/APosition.java
@@ -10,7 +10,7 @@ public abstract class APosition {
 
     public abstract void addVector(List<Integer> dimension);
 
-    public abstract float getyCoordinate();
+    public abstract int getyCoordinate();
 
-    public abstract float getxCoordinate();
+    public abstract int getxCoordinate();
 }
